It's flying!
After three days of constant PID tuning, it finaly does as expected. The
sensitivity of the radio controller input have been significantly
reduced. Variable pitch seems to give fast change, and the standard
input is way to high for a pilot to control the quadrotor. The P gain
for the pitch and roll angle have been increased by a factor of almost
seven!